krisna_experiments_codes/requirements.txt
torch
transformers
sentence_transformers
datasets
scipy
plotly-express
matplotlib
numpy
einops